1. support post id 2. dry run flag 3. fix line break bug

+++ download.js
@@ -7,9 +7,11 @@ var htmlEscape = require("html-escape");
 var getOpt = require('node-getopt');
 
 var opt = getOpt.create([
+    ['', 'post[=]' , 'post id'],
     ['', 'page[=PAGE]'  , 'Rnages of pages to fetch (start:end)'],
     ['', 'output[=DIR]' , 'Output'],
     ['', 'overwrite'    , 'Whether to overwrite existing chapters'],
+    ['', 'dryrun'       , 'dry run'],
     ['h', 'help'        , 'Display help']
 ]).bindHelp().parseSystem();
 
@@ -41,13 +43,15 @@ function writeFile(fileName, text) {
       if (opt.options['overwrite']) {
         console.log('Overwrite', fileName);
       } else {
-        console.log('Skip', fileName); 
+        console.log('Skip', fileName);
         return;
       }
     } else {
       console.log('New file', fileName);
     }
-    fs.writeFile(fileName, text);
+    if (!opt.options['dryrun']) {
+      fs.writeFile(fileName, text);
+    }
   });
 }
 
@@ -58,17 +62,17 @@ function processSections(index, sections) {
     node.find('.pstatus').remove();
     node.find('ignore_js_op').remove();
     node.find('a').remove();
-    node.find('br').replaceWith('\n\n');
+    node.find('br').replaceWith('\n');
 
     // Normalize line break and space.
     // Remove bad char which causes problems when translating to epub (0x0c).
     var text = node.text()
-      .replace(/\r\n/g, '').replace(/\u3000|\xa0/g, ' ')
+      .replace(/\r\n/g, '\n').replace(/\u3000|\xa0/g, ' ')
       .replace(/\x0c/g, '');
-     
+
     // epub creator often treats text as html so needs to escape.
     text = htmlEscape(text);
-    
+
     // Try best to guess redundant line break used to control layout on forum.
     text = text.replace(/([^。,？！」])\n\n([^\s])/g, '$1$2');
 
@@ -78,13 +82,11 @@ function processSections(index, sections) {
   return index;
 }
 
-var URL_TEMPLATE = 'http://ck101.com/thread-1586268-%s-1.html';
-
-function download(startPage, endPage) {
+function download(post_url, startPage, endPage) {
   // Use promise to chain parallel responses in order.
   var all = Promise.resolve(1);
   for (var i = startPage; i <= endPage; ++i) {
-    var url = util.format(URL_TEMPLATE, i);
+    var url = util.format(post_url, i);
 
     // Create a closure so the "page" is copied. Otherwise the same
     // object is referenced when page is done in each iteration.
@@ -98,20 +100,23 @@ function download(startPage, endPage) {
   }
 
   all.then(function() {
-    console.log('All Done.');    
+    console.log('All Done.');
   }, function(err) {
     console.log(err);
   });
 }
 
-function getPageNum() {
+function getPageNum(post_url) {
   console.log('Getting num of pages');
-  var firstPage = util.format(URL_TEMPLATE, 1);
+  var firstPage = util.format(post_url, 1);
   return getOnePage(firstPage).then(function(content) {
     $ = cheerio.load(content);
     return parseInt($('.pgt .pg a.last').text().match(/\d+/)[0], 10);
   });
 }
+
+var YAODAO_ID = '1586268';
+var URL_TEMPLATE = 'http://ck101.com/thread-%s-%s-1.html';
 
 function main() {
   if ('output' in opt.options) {
@@ -123,6 +128,9 @@ function main() {
     };
   }
 
+  console.log(opt.options['post']);
+  var post_url = util.format(URL_TEMPLATE, opt.options['post'] || YAODAO_ID);
+
   var startPage = 1;
   var endPage = null;
   if ('page' in opt.options) {
@@ -131,10 +139,10 @@ function main() {
     endPage = parseInt(pageRange[1], 10) || endPage;
   }
 
-  var last = endPage ? Promise.resolve(endPage) : getPageNum();
+  var last = endPage ? Promise.resolve(endPage) : getPageNum(post_url);
   last.then(function(endPage) {
     console.log('start page', startPage, 'end page', endPage);
-    download(startPage, endPage);
+    download(post_url, startPage, endPage);
   });
 }
